+++ src/Infrastructure/Persistence/DataModels/HumanResources/EmployeePayHistory.cs
@@ -5,7 +5,7 @@ namespace REA.Accounting.Infrastructure.Persistence.DataModels.HumanResources
         public int BusinessEntityID { get; set; }
         public DateTime RateChangeDate { get; set; }
         public decimal Rate { get; set; }
-        public int PayFrequency { get; set; }
+        public byte PayFrequency { get; set; }
         public DateTime ModifiedDate { get; set; }
     }
 }
